// src/types/PropertiesInfo.ts
export type FrontCardProps = {
  urlImage: string;
  aluguel: number;
  cidadeEstado: string;
  bairro: string;
  dormitorio: number;
  banheiro: number;
  vagas: number;
  areaPrivativa: number;
}

export type BackCardProps = {
  aluguel: number;
  condominio: number;
  iptu: number;
  areaTotal: number;
  areaPrivativa: number;
  sobreImovel: string;
}